refactor(editor): Move useStyles into @n8n/composables behind a shim (no-changelog)
